created crud operations with news

+++ app/src/main/java/ru/clevertec/news/repositories/NewsRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.clevertec.news.entities.news.News;
 
 @Repository
-public interface NewsRepository extends JpaRepository<News, Long> {
+public interface NewsRepository extends JpaRepository<News, Integer> {
 }
